Share `Tag2` impl between `CopyTaggedPtr` and `TaggedPtr` tests
